requirements.txt
alpaca-py
yfinance~=1.2.0
numpy~=2.4.3
pandas~=3.0.1
dotenv~=0.9.9
python-dotenv~=1.2.2